fix type class resolution and plugin publishing, still need to autoinject compiler free args
