+++ gradle-plugin/build.gradle
@@ -25,7 +25,7 @@ compileTestKotlin {
 
 jar {
   // Embed compiler plugin in jar
-  configurations.named("compileClasspath").get().find { it.name.startsWith("plugin-compiler") }
+  configurations.named("compileClasspath").get().find { it.name.startsWith("compiler-plugin") }
 
   manifest {
     attributes["Specification-Title"] = project.name
